load_data/py_requirements.txt
pandas==2.0.3
redshift-connector==2.0.914
loguru==0.7.2
PyYAML==6.0.1